refactor: consts.ts -> const.ts
